Create code update (active-red)
This code update is for assessing better interpretation of the results from the PAA analysis. It works on the same principle as the previous version with a single exception: the final png is not a "black & white" format, but the active parts of the video are highlighted as "red" and the inactive background is derived from the first frame of the video.